Fix erroneous missing log4j error.
To facilitate more control over the logging framework, Logstash manually instantiates the log4j2 context very early in the lifecycle. However, BEFORE we can instantiate our own log4j context, Java loggers in static blocks (very, very early in the lifecycle) results in a default context getting initialized. The process of initializing the default context results in error message stating that log4j2.properties could not be found. This means that 2 contexts are in use, one very, very early in the lifecycle (Java static initialization),  and other context explicitly established and used.

This change is to configure the first context by simply shipping a default log4j2.properties on the class path that only log errors to the console. This is functionally equivalent to prior behavior, but will not show the erroneous error message.